Moved Cursor functions to new Cursor class
