+++ src/KeyboardController.ts
@@ -6,12 +6,14 @@ import { Constant } from "./Constant";
 import App from "./App";
 import { copyTowerObject } from "./CopyTowerObject";
 import { Brick } from "./Brick";
+import Cursor from "./Cursor";
 
 const log = console.log;
 
 export default class KeyboardController {
   public history: History;
   private app: App;
+  private cursor: Cursor;
 
   constructor(app: App) {
     this.app = app;
@@ -22,93 +24,6 @@ export default class KeyboardController {
     document.addEventListener("keydown", e => {
       this.onKeyDown(e);
       return true;
-    });
-  }
-
-  public findCanAboveCursor() {
-    const result = findById(
-      this.app.currentTower(),
-      this.app.state.canCursorId
-    );
-    if (!result) {
-      log("No result found. Assuming bottom.");
-      return this.app.currentTower().rootBrick.uniqueId;
-    }
-    const brick = result.brick;
-    if (brick instanceof Invocation && brick.args.length > 0) {
-      return brick.args[0].uniqueId;
-    } else {
-      return null;
-    }
-  }
-
-  public findCanBelowCursor() {
-    const result = findById(
-      this.app.currentTower(),
-      this.app.state.canCursorId
-    );
-    console.log("Trying to find below cursor", result);
-    if (!result) {
-      log("No result found. Assuming bottom.");
-      return this.app.currentTower().rootBrick.uniqueId;
-    }
-    if (result.path.length === 0) {
-      log("Bottom of tower. Cannot go down.");
-      return null;
-    }
-    return result.path[result.path.length - 1].uniqueId;
-  }
-
-  public findCanToLeftOfCursor() {
-    const result = findById(
-      this.app.currentTower(),
-      this.app.state.canCursorId
-    );
-    if (!result) {
-      log("No result found. Assuming bottom.");
-      return this.app.currentTower().rootBrick.uniqueId;
-    }
-    if (result.path.length === 0) {
-      log("Bottom of tower. Cannot go down.");
-      return null;
-    }
-    const parent = result.path[result.path.length - 1];
-    const index = parent.args.indexOf(result.brick);
-    if (index === 0) {
-      // just move down.
-      return parent.uniqueId;
-    }
-    return parent.args[index - 1].uniqueId;
-  }
-
-  public findCanToRightOfCursor() {
-    const result = findById(
-      this.app.currentTower(),
-      this.app.state.canCursorId
-    );
-    if (!result) {
-      log("No result found. Assuming bottom.");
-      return this.app.currentTower().rootBrick.uniqueId;
-    }
-    if (result.path.length === 0) {
-      log("Bottom of tower. Cannot go down.");
-      return null;
-    }
-    const parent = result.path[result.path.length - 1];
-    const index = parent.args.indexOf(result.brick);
-    if (index + 1 === parent.args.length) {
-      // just move down.
-      return parent.uniqueId;
-    }
-    return parent.args[index + 1].uniqueId;
-  }
-
-  public moveCursorTo(uniqueId: string | null) {
-    if (uniqueId === null) {
-      return;
-    }
-    this.app.setState({
-      canCursorId: uniqueId
     });
   }
 
@@ -310,16 +225,16 @@ export default class KeyboardController {
         this.editConstant();
         break;
       case "KeyJ":
-        this.moveCursorTo(this.findCanBelowCursor());
+        this.cursor.moveCursorTo(this.cursor.findCanBelowCursor());
         break;
       case "KeyK":
-        this.moveCursorTo(this.findCanAboveCursor());
+        this.cursor.moveCursorTo(this.cursor.findCanAboveCursor());
         break;
       case "KeyH":
-        this.moveCursorTo(this.findCanToLeftOfCursor());
+        this.cursor.moveCursorTo(this.cursor.findCanToLeftOfCursor());
         break;
       case "KeyL":
-        this.moveCursorTo(this.findCanToRightOfCursor());
+        this.cursor.moveCursorTo(this.cursor.findCanToRightOfCursor());
         break;
       case "KeyC":
         this.copyToSky();
